+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Validate on new exim id format (#1294)
 - Permit + and = chars in email for sendmail (#1291)
 - Fix Message Listing on Mysql 8+ (#1295)
+- Fix MaxMind database download when curl extension is not available (#1297)
 
 ## 1.2.21
 ### Improvements